--- handler.js
@@ -1,2 +1,0 @@
-export * from "./product-service/get-product-list";
-export * from "./product-service/get-product-by-id";
