+++ product-service/handler.js
@@ -1,0 +1,2 @@
+export * from "./products/get-product-list";
+export * from "./products/get-product-by-id";
