+++ product-service/webpack.config.js
